Streamlines Builder setup
