+++ flake-ctl/flake-ctl-build/src/lib.rs
@@ -1,11 +1,10 @@
-use std::{fs::remove_dir_all, io::stdin, env::var, process::Command};
 pub use std::path::{Path, PathBuf};
+use std::{env::var, fs::remove_dir_all, io::stdin, process::Command};
 
-use anyhow::{Result, anyhow, Context};
-use clap::Args;
-use flakes::config::{itf::FlakeConfig, self};
+use anyhow::{anyhow, Context, Result};
+use clap::{Args, FromArgMatches, Parser};
+use flakes::config::{self, itf::FlakeConfig};
 use tempfile::tempdir;
-
 
 pub trait FlakeBuilder {
     /// Create the infrastructure to build the flake package in the given directory
@@ -20,23 +19,23 @@ pub trait FlakeBuilder {
     fn build(&self, options: &PackageOptions, target: Option<&Path>, location: &Path) -> Result<()>;
 
     /// Remove all the temporary infrastructure created in [setup]
-    /// 
+    ///
     /// This may be run before _and_ after each build
-    /// 
+    ///
     /// Only modify the contents of `location`, if the directory itself needs to be deleted this will be handled by [cleanup_default_directory]
     fn cleanup(&self, location: &Path) -> Result<()>;
 
     /// What directory to use when no build location is given, this defaults to a new tmp directory
-    /// 
+    ///
     /// Override this for your builder if there is another sensible default
-    /// 
+    ///
     /// Keep in mind that you may also have to adjust [cleanup_default_directory]
     fn get_default_build_directory(&self) -> Result<PathBuf> {
         Ok(tempdir()?.into_path())
     }
 
     /// Cleanup the directory structure of the default directory, this is only called when no build location was given
-    /// 
+    ///
     /// Default behaviour is to just delete the directory (assumed to be a tmp directory created by [get_default_build_directory]),
     /// override this function if this is undesired
     fn cleanup_default_directory(&self, dir: &Path) -> Result<()> {
@@ -47,31 +46,28 @@ pub trait FlakeBuilder {
     ///
     /// If `target` is given store the resulting package there, otherwise put it into the current working directory
     /// If `location` is given the package is build there, otherwise it is constructed in a tmp directory
-    /// 
+    ///
     /// [build] is only run if `build` is `true`
-    /// 
+    ///
     /// [cleanup] is not run if `keep` is `true`
     fn execute(
-        &self, options: &PackageOptions, target: Option<impl AsRef<Path>>, location: Option<impl AsRef<Path>>, build: bool, keep: bool,
+        &self, options: &PackageOptions, target: Option<impl AsRef<Path>>, location: Option<impl AsRef<Path>>, build: bool,
+        keep: bool,
     ) -> Result<()> {
         let cleanup_default = location.is_none();
-        
+
         let location = match location.as_ref() {
             Some(location) => location.as_ref().to_owned(),
             None => self.get_default_build_directory()?,
         };
-        
+
         // Clean the location if it is reused, ignore errors here
         self.cleanup(&location).ok();
         self.setup(&location)?;
         let config = config::load_from_target(Path::new(&options.name))?;
         self.create_bundle(options, &config, &location)?;
 
-        let result = if build {
-            self.build(options, target.as_ref().map(AsRef::as_ref), &location)
-        } else {
-            Ok(())
-        };
+        let result = if build { self.build(options, target.as_ref().map(AsRef::as_ref), &location) } else { Ok(()) };
 
         if !keep {
             self.cleanup(&location)?;
@@ -86,21 +82,20 @@ pub trait FlakeBuilder {
         let options = args.determine_options()?;
         self.execute(&options, args.target.as_ref(), args.location.as_ref(), !args.dry_run, args.keep)
     }
-
 }
 
 #[derive(Debug)]
-pub struct PackageOptions{
+pub struct PackageOptions {
     pub name: String,
     pub description: String,
     pub version: String,
     pub url: String,
     pub maintainer: String,
-    pub license: String
+    pub license: String,
 }
 
 #[derive(Debug, Default, Args, Clone)]
-pub struct PackageOptionsBuilder{
+pub struct PackageOptionsBuilder {
     #[arg(long)]
     pub name: Option<String>,
     #[arg(long)]
@@ -112,18 +107,18 @@ pub struct PackageOptionsBuilder{
     #[arg(long)]
     pub maintainer: Option<String>,
     #[arg(long)]
-    pub license: Option<String>
+    pub license: Option<String>,
 }
 
 impl PackageOptionsBuilder {
     pub fn build(self) -> Result<PackageOptions> {
         Ok(PackageOptions {
-            name: self.name.ok_or_else(||anyhow!("Missing package name"))?,
-            description: self.description.ok_or_else(||anyhow!("Missing package description"))?,
-            version: self.version.ok_or_else(||anyhow!("Missing package version"))?,
-            url: self.url.ok_or_else(||anyhow!("Missing package url"))?,
-            maintainer: self.maintainer.ok_or_else(||anyhow!("Missing package maintainer"))?,
-            license: self.license.ok_or_else(||anyhow!("Missing package license"))?,
+            name: self.name.ok_or_else(|| anyhow!("Missing package name"))?,
+            description: self.description.ok_or_else(|| anyhow!("Missing package description"))?,
+            version: self.version.ok_or_else(|| anyhow!("Missing package version"))?,
+            url: self.url.ok_or_else(|| anyhow!("Missing package url"))?,
+            maintainer: self.maintainer.ok_or_else(|| anyhow!("Missing package maintainer"))?,
+            license: self.license.ok_or_else(|| anyhow!("Missing package license"))?,
         })
     }
 }
@@ -184,10 +179,9 @@ pub struct BuilderArgs {
 }
 
 impl BuilderArgs {
-
     pub fn determine_options(&self) -> Result<PackageOptions> {
         let mut options = self.options.clone();
-        
+
         // Read from env where not given
         options.name = options.name.or_else(|| var("PKG_FLAKE_NAME").ok());
         options.description = options.description.or_else(|| var("PKG_FLAKE_DESCRIPTION").ok());
@@ -195,7 +189,7 @@ impl BuilderArgs {
         options.url = options.url.or_else(|| var("PKG_FLAKE_URL").ok());
         options.maintainer = options.maintainer.or_else(|| var("PKG_FLAKE_MAINTAINER").ok());
         options.license = options.license.or_else(|| var("PKG_FLAKE_LICENSE").ok());
-    
+
         if !self.ci {
             options.name = options.name.or_else(|| user_input("Name").ok());
             options.description = options.description.or_else(|| user_input("Description").ok());
@@ -204,21 +198,33 @@ impl BuilderArgs {
             options.maintainer = options.maintainer.or_else(|| user_input("Maintainer").ok());
             options.license = options.license.or_else(|| user_input("License").ok());
         }
-    
+
         options.build().context("Missing packaging option")
     }
-    
-    }
-    
-    fn user_input(name: &str) -> Result<String> {
-        let mut buf = String::new();
-        println!("{name}: ");
-        stdin().read_line(&mut buf)?;
-        Ok(buf.trim_end().to_owned())
-    }
+}
 
+fn user_input(name: &str) -> Result<String> {
+    let mut buf = String::new();
+    println!("{name}: ");
+    stdin().read_line(&mut buf)?;
+    Ok(buf.trim_end().to_owned())
+}
 
-    pub fn export_flake(name: &str, pilot: &str, bundling_dir: &Path) -> Result<()> {
-        Command::new("flake-ctl").arg(pilot).arg("export").arg(name).arg(bundling_dir.join(name)).status()?;
-        Ok(())
-    }
+pub fn export_flake(name: &str, pilot: &str, bundling_dir: &Path) -> Result<()> {
+    Command::new("flake-ctl").arg(pilot).arg("export").arg(name).arg(bundling_dir.join(name)).status()?;
+    Ok(())
+}
+
+#[derive(Parser)]
+struct FullArgs<T: FromArgMatches + Args + FlakeBuilder> {
+    #[command(flatten)]
+    args: BuilderArgs,
+
+    #[command(flatten)]
+    builder: T,
+}
+
+pub fn run<T: FromArgMatches + Args + FlakeBuilder>() -> Result<()> {
+    let command = FullArgs::<T>::parse();
+    command.builder.run(&command.args)
+}
